src/main/java/Formatter.java
import java.io.*;

/**
 * Класс Formatter предназначен для форматирования текста
 */
 class Formatter implements IFormatter {
   public Formatter(){
    }

    /**
     *метод format получает путь к файлу и форматирует его содержимое
     * @param in входные данные
     * @param out выходные данные
     * @throws IOException исключение
     */
    public  String format(final IReader in, final IWriter out) throws IOException {
        int indent = 0;
        String line;
        String newString = "";
        String space = "";
        while ((line = in.readCharline()) != null) {
            line = line.trim();
            char[] buffer = new char[line.length()];
            buffer = line.toCharArray();
            for (int i = 0; i < buffer.length; i++) {

                if (!(buffer[i] == '{' | buffer[i] == '}' | buffer[i] == ';' | buffer[i] == buffer.length - 1)) {
                    newString = newString + buffer[i];
                } else {
                    if (buffer[i] == '{') {
                        indent = indent + 4;
                        space = "";
                        for (int a = indent; a > 0; a--) {
                            space = space + " ";
                        }
                        newString = newString + buffer[i] + "\r\n" + space;
                    }
                    if (buffer[i] == '}') {
                        indent = indent - 4;
                        space = "";
                        for (int a = indent; a > 0; a--) {
                            space = space + " ";
                        }
                        newString = newString + "\r\n" + space + buffer[i] + "\r\n" + space;
                    }
                    if (buffer[i] == ';') {
                        newString = newString + buffer[i] + "\r\n" + space;
                    }
                }
            }

        }
        out.writeChars(newString);
        out.close();



return newString;
    }


}

